[dPWA/Link Capturing] Move supported_links items to cr_components
This CL moves the supported_links item component and the overlapping
apps dialog frontend component to cr_components/ so that it can be
reused by both Desktop and CrOS platforms.

The ReduxStore based behavior is removed, and the sites that show
the supported links item instead read the apps from the ReduxStore
and pass it to the components via Polymer 1 way binding.

Verified manually that this works, need to fix the OSSettingsTest for
this component.

Bug: 1412856, b/294232431
Change-Id: Ie7774f7eb674eeaf3a0841a24e31dffce227ca26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4749623
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1180889}