Check for non-zero selection bounds when showing touch selection menu.
Prevent the aura touch selection menu from appearing when the selection
bounds are zero, since this usually means that touch selection is not
active or that there is no cursor or selection. Currently in these
cases, a stray menu ends up incorrectly appearing at the top left of the
screen/app/content while the touch handles are not shown. Since the
touch handles weren't shown, the menu probably shouldn't have been shown
either.

Need to further investigate why the menu is being triggered in these
situations (seems that it is handling a long press context menu event
which it shouldn't handle), but for now just prevent the stray menu from
appearing. This is similar to how empty selection bounds are handled in
views so I think it should be ok:
https://source.chromium.org/chromium/chromium/src/+/main:ui/views/touchui/touch_selection_controller_impl.cc;l=739;drc=8f8154598d766545ee1e8234f93a895659f62434

(Also fix a test so that it waits for selection updates to occur before
checking whether the menu is shown, since this change means that the
menu won't be shown if no caret or selection has been set.)

Bug: 1470581, b:286348351
Change-Id: Ieae03fe5cd615ebc8c70b42382c6dcf28ca83f7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4755023
Commit-Queue: Michelle Chen <michellegc@google.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1180896}