[android][trackpad] Allow Trackpad Gesture Improvements for Android U+
For the track-pad gesture improvements we need certain signals from the
MotionEvent that are only available on Android U+. Therefore we need to
ensure that devices running Android versions < Android U (34) do not run
the code that converts the track-pad MotionEvents to mouse events.

The code to do this was originally implemented in this CL -
https://chromium-review.googlesource.com/c/chromium/src/+/4355011.

Because we will be enabling the track-pad gesture improvements in M117,
this CL will need to be cherry-picked into that milestone.

Bug: 1471044
Change-Id: I519dd6ae6db2f6d55f4815924aeab6d2594f293d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4758527
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Wayne Jackson Jr. <wbjacksonjr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1180918}