Send line bounds through CursorAnchorInfo on requestCursorUpdate
The coordinates are calculated in Blink and ImeCompositionRangeChanged.
These coordinates are in the document coordinate space. They're sent
through Mojo and JNI to the CursorAnchorInfoController where the
current page's zoom and scroll offsets are applied to the coordinates.
After the transform, the new coordinates are attached to the CursorAnchorInfo object and sent to Android.

Bug: 1442938
Change-Id: Ib112f8cb9daea322207f2271c5be1c2535bee3da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4394588
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Alex Mitra <alexmitra@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1180994}