Capture paint preview of all subframes in prerendered content
PaintPreviewClient gets stuck and never invokes completion callback
when it is run on prerendered pages with subframes.
http://www.chromestatus.com/metrics/css/timeline/popularity/4 is an
example of such a page. The reason is that paint preview is never
called for the subframes in prerendered content and therefore the
`awaiting_subframes` member of InProgressDocumentCaptureState never
gets empty. This causes the following if condition to always fail in
PaintPreviewClient::OnPaintPrewviewCaptureCallback and the completion
callback is never fired:

  if (document_data->awaiting_subframes.empty())
    OnFinished(params.document_guid, document_data);


The fix is to make NoStatePrefetchContents override
WebContentsDelegateImpl::CapturePaintPreviewOSubframe and ensure that
PaintPreviewClient is called for each subframe. This is similar to what
other delegates such as Browser::CapturePaintPreviewOfSubframe do when
paint preview is run on the same page when it is navigated to via the
address bar and is not prerendered. Here is the callstack for this
scenario:

Browser: :CapturePaintPreviewOfSubframe
content: :WebContentsImpl::CapturePaintPreviewOfCrossProcessSubframe
content: :RenderFrameProxyHost::CapturePaintPreviewOfCrossProcessSubframe
Change-Id: I748d9d62606160c64da86e4ad5c54347d96b54c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4757283
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1180997}