webcodecs: Split frame into horizontal and vertical tiles in AV1 encoder
This CL tweaks the AV1 encoder so that the frame is split into both rows and columns (instead of just columns).
Tests for webrtc showed that it improves encoding perf and there is no reason to think that it will be different for webcodecs.

Change-Id: If30c3d1a4f7cf83479aba8e31d2ae091441bafd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4757920
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181030}