[autopip] [mediasession] Use top frame's service if it exists
Currently, if no frames of a WebContents have requested audio focus,
then no media session actions can be routed (since the MediaSessionImpl
decides that no frame's service should be the routed service). This
presents an issue for some video conferencing services where media
session action handlers are needed. In the past this was an issue for
video-conferencing-specific media session actions, but since those were
only used in picture-in-picture we just had an exception for PiP
players. However, the new 'enterpictureinpicture' action is a useful
action for video conferencing websites (and can only be used outside of
PiP), so we require a different solution.

This CL changes the media session service routing logic to allow the
topmost frame to be routed even if it hasn't requested audio focus as
long as no other frames would otherwise be routed. Note that this won't
mean that the frames will start requesting audio focus, just that they
can receive media session actions.

Bug: 1457056
Change-Id: Ib1a7eb61e6cf4abf1d5fe5396229c6401f6074e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4659151
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181055}