[Tailored security] add a preference to store the next sync flow time
We're simplifying the retry feature of the sync flow of tailored
security to use a single retry time value. If the retry state has never
been set, this pref value will be set to far in the future, example: 90
days. If at some point the sync flow starts, then this value will be
updated to 1 day from now so that the retry mechanisms can try again
soon.

Change-Id: Id387127481caf479b8e8cd257f25f3167321dd08
Bug: 1470916
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4757643
Commit-Queue: Javier Castro <jacastro@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181057}