[ntp][drive] Retrieve segmentation result before showing Drive module
Retrieves prediction (classification result) from segmentation plaform before showing drive module (when the NtpDriveModuleSegmentation experiment is enabled).

With this CL, the classification result will be ignored. Future CLs will make use of the result to conditionally show (or not show) the Drive module.

Making a call to the segmentation platform before showing Drive files allows the platform to get better signals for when/how a user will use Drive. For example, if the API is called and returns 'show_drive' then the user does not interact with the module, future predictions can account for this.

By getting this into M117, we can bolster the reliability of our classification results, before using them in future milestones, to show or not show the Drive module.


Change-Id: If68f41096ca6998f2b0a16d10b0914962a9a6974
Bug: 1470762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4754911
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Siddhartha S <ssid@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Paul Adedeji <pauladedeji@google.com>
Cr-Commit-Position: refs/heads/main@{#1181065}