Fixed the crash that the SearchResultImageView has empty a11y name
As title. This is because the image view's a11y name is not set if the
result update is run but no image is set. Fixed that by setting the
a11y name no matter if the image is set or not. This makes sense as
the a11y functionality doesn't really depend on if the image is shown
or not.

Bug: 1352636
Change-Id: If38f21f63e3556f18f32acef7d37b8f2330e2806
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4760253
Reviewed-by: Yulun Wu <yulunwu@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181067}