Display the details of PDF attachment(s)
Add `viewer-attachment-bar` and `viewer-attachment` for displaying the
details of PDF attachments:

`viewer-attachment`:
Display an attachment's name and provide a button for download.
`viewer-attachment-bar`:
Contain a list of `viewer-attachment`. If any of the attachment
exceeds 100MB, display a `warning` that such attachment is not
available for download.

Bug: 177188
Change-Id: I35cbe5624a49db0bc5493fc51d3ff4cc90b41cdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4712197
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Nigi <nigi@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1181091}