Mac immersive: handle events while showing alert in fullscreen
Currently, windows that should not handle their own commands defer to
a "bubble parent", defined as a parent window which conforms to
`CommandDispatchingWindow`.

Unfortunately, all `NativeWidgetMacNSWindow` including the overlay
window conform to this protocol. That means when an alert is showing
in full-screen, its bubble parent is the overlay window which can't
handle events by itself, but does not itself bubble to the browser
window as it should, since it's in the toolbar's hierarchy.

This change pushes the bubble parent logic out of command dispatcher
and into `NativeWidgetMacNSWindow` (and renames it to
`commandDispatchParent` to minimize confusion with "bubble" as a noun).
This allows us to set an explicit override; we take advantage of this to
set the browser window as the overlay window's command dispatch parent.

Bug: 1469896
Change-Id: Ie8d68a018683881c7e63d63c87e8f77e07d71b93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4756745
Reviewed-by: Jayson Adams <shrike@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1181103}