[A11y] Fix for longstanding hypertext errors
Two root causes were found and fixed:
1. An iframe or other child tree that is destroyed must update the
hypertext of the owning node.
2. A node that becomes a leaf must still update its hypertext.
Previously we did not add that previous hypertext owning parent.

Also required fixing of firing of alert event in
content_setting_image_view.cc. It fired a redundant event, and the alert
was not fired on an object with the correct role, which triggers a DCHECK unrelated to the problem solved in this CL. The fix is to just set the correct role. This change was required in order to get the new page tab a11y browsertests to pass. Alerts work on every platform, so there is no need for the #if.

Fixed: 1404671
Change-Id: Id7f0624e98a8577dd4656f24ac63972adc5a38eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4753355
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1181145}