FSA: Implement method to retrieve dormant persisted grants
Implements `GetDormantPersistedObjects` which returns a set of the
dormant grants for a given origin. This method is needed to support
FSA Persistent Permissions implementation, and will be used in followup
changes to implement the restore prompt feature.

Implements helper method `HasGrantedActiveGrant` whose return
value is one of the determining factors in retrieving dormant grants.

Updates naming of `GetExtendedGrantedObjects` ->
`GetExtendedPersistedObjects` for naming consistency.

Adds a boolean parameter to the
`SetOriginHasExtendedPermissionForTesting` method to allow for
testing dormant grants scenarios.

Bug: 1011533
Change-Id: Ic7c21325fce46c41580d7347f296f49e2133f911
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4752712
Reviewed-by: Austin Sullivan <asully@chromium.org>
Commit-Queue: Christine Smith <christinesm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181188}