[ios] Add a unit test for NTPCoordinator's selectFeedType
Adds a unit tests to verify that the correct view controller is loaded
when the following feed and the discover feed are selected.

In order to add this test, TestingApplicationContext had to be changed
to allow setting a VariationsService, so that the stored permanent
country could be overridden.

Bug: 1464692
Change-Id: Ia613552e58ed128c19e2b4c6ab34a7732dd1d4c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4721591
Reviewed-by: Adam Arcaro <adamta@google.com>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181281}