[Cookies] Add support for blocking truncated cookies
Adds the code needed to block truncated cookies behind a feature
flag (currently disabled). This also plumbs the value such that
this can be controlled by an enterprise policy.

This is part of a chain of CLs updating the cookie handling logic to
ignore truncated cookies. The following prototype CL shows all of the
planned changes in one place:
https://chromium-review.googlesource.com/c/chromium/src/+/4134025

Bug: 1233602
Change-Id: I513f5a613e9260e63bd4be2cf26ca60aa1c273f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4706696
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Reviewed-by: Nick Harper <nharper@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181302}