[Extensions Cleanup] Move CanShowActionsInToolbar to ToolbarModel
CanShowActionsInToolbar was computed in the extensions container because
that was the only view that needed that value. However,
chrome://extensions also need to know if an extension can be shown
in the toolbar. Therefore, we can move this to the ToolbarModel which
arguably is a better place since we are computing a value for actions.

Bug: N/A
Change-Id: I0341b10dc154aff8b574eabb1475a9de8709ff25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4740660
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181305}