Added an ApplyAutofillAction function in the renderer
This CL:
-Adds a mojom::AutofillFormAction enum denoting the browser autofill
operation that sent a form to the renderer.
-Adds an ApplyAutofillAction function to form_util with intent to merge
the two deleted functions.
-Adapts FillFormField and PreviewFormField and ClearPreviewedElements
to handle both Fill and Undo operations.
-Fixes typos and clang warnings in touched files

The overall chain of CLs adds a preview feature to the existing Undo
functionality. Up until this CL, the browser has the required logic and
sends the necessary information to the renderer. This CL's goal is to
first make Undo previewable and second merge the Undo and Fill functions
in the renderer because they are highly similar.

Bug: 1441410
Change-Id: Id2ce96d4df2318469a2a2578257040840009e30c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4664652
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1181334}