[iOS] Add custom table view to family picker VC
Initial implementation of FamilyPickerVC was a subclass of
TableViewBottomSheetViewController. This CL replaces this inheritance
with a custom table view implementation as the inheritance comes with
a lot of custom logic that is not needed for this view.

Screen after modification: http://screen/7tL4j3jf5dQjrqy

Bug: 1463882
Change-Id: I984c064d2a13107cb10bcf85094f1689d40d0154
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4755463
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Rafał Godlewski <rgod@google.com>
Cr-Commit-Position: refs/heads/main@{#1181401}