[SAA][Settings] Fix keyboard navigation
This CL fixes the keyboard navigation of the Storage Access settings
page. We utilize the FocusGrid to construct a grid of elements for
keyboard navigation including arrows. We do not use the FocusRowMix
because it cannot find focus-row-controls on the nested elements,
which we want.

Demo with arrows: http://dr/file/d/1_Y-vej8Tjoex45LVup-PixlfiH_3ZUkX/view

Fixed: b:294355587
Bug: 1433644, 1471396
Change-Id: I40a3d604ce3a42e4bb5e5b038aa04f37afdfe621
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4757381
Commit-Queue: Filipa Senra <fsenra@google.com>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#1181407}