[autofill autocomplete] Remove mouse observed outside bounds check for autocomplete being deleted case.
When an autocomplete entry is deleted the whole popup is recreated. This leads to the cursor being inside a row from the beginning. Due to our mouse being outside the bounds check, this forces the user to have to move the cursor out and in, which is annoying.

This code also deleted DeleteButtonAccessibilityDelegate which is no longer used, plus renames the method to create the autocomplete row.

please note: There is still a bug where a mouse movement is required to get the cell highlighted. I am working on it.

https://screencast.googleplex.com/cast/NTI0MDcwMjA1MjA3MzQ3Mnw1YmNkMDYwYS1jMQ

Bug: 1417187
Change-Id: I1e1dd4ed41dc05dbb288d05d46e89a332d0aaab4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4758525
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Bruno Braga <brunobraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1181419}