[iOS] Allow Spacing Increase If Info Button Is Hidden
TableViewInfoButtonItem has functionality to hide the info button, but
the constraint wasn't properly set up in a way where the UI looked like
space was added for a button that was invisible. Please see UI
screenshot for reference. Although this is what happens visually, what
actually is causing this UI bug is the logic that space is added for
`statusTextLabel` even if it doesn't exist. To fix this, we separate
the constraints related to the UILabel and activate/deactivate related
constraints based on if the UILabel was initialized.

UI Screenshots:
Standard:
https://drive.google.com/file/d/1BiVEYgonMyHByjVoXNi8X2a17608Azc-/view?usp=drive_link&resourcekey=0-0URX4AvIJUOIstkYBBoODA

Accessibility Enabled:
https://drive.google.com/file/d/1TFJVUzOADdgrqnniPSRoySwqaGy1Rj5s/view?usp=drive_link&resourcekey=0-YwQEaToOjxlW8vRGK3euQQ

Bug: 1470372
Change-Id: I3c1bd9a874d0b060ef99a65e907dcbdf2aebd43e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4753894
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181441}