[Bookmarks] client_ gets destructed before the observer list.
Before this CL, the observer list was destructed before client_.
Since `client_` owns a BookmarkClient and some client (at least
`ChromeBookmarkClient`) observer the bookmark model, it is
important that the client is destructed before the list. This caused
the destruction of a non-empty list.

While reordering parameters would suffice to ensure the right order,
adding the destruction of the client explicitly in the destructor of
the model ensure that this requirement is explicitly met, even if
members of the class were to be reordered.

`TutledUrlIndex` owns  a `TypedCountSorter` that keeps a raw_ptr to the client. So `titled_url_index_` must be reset before the `client_`.
Hence this CL also had to explicitly stops `titled_url_index_` before
`client_`.

There may remain more dependencies in the order of destructor that are
not explicited in comment. The goal of this CL was not to make them
all explicit, only to ensure the observer list contains less elements
at destruction time.



This CL will allow to check it on observer list destruction in
crrev.com/c/4742500.

Bug: None
Change-Id: Icaeee162580848829ff83861d1ed561df929de1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4755448
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181469}